+++ .editorconfig
@@ -14,7 +14,7 @@ root = true
 [*]
 charset = utf-8
 indent_style = space
-indent_size = 2
+indent_size = 1
 end_of_line = lf
 trim_trailing_whitespace = true
 insert_final_newline = true
